mata: Declare PRODUCT_SOONG_NAMESPACES for HALs
HALs are starting to move to blueprint files, which means that pathmap
ifeqs no longer work. This results in modules with the same name in
different platform HAL directories. Since these modules aren't used or
linked anywhere outside the HAL directories and a device's device.mk,
we can safely use a soong namespace to isolate the different HAL
directories' blueprints. This defines the PRODUCT_SOONG_NAMESPACES
make flag so make knows which namespace to look in as well.

Wight554: Adapt for our pathmap

Change-Id: I549e6da863465d79b12c7453aa169d5d963afa52
Signed-off-by: celtare21 <celtare21@gmail.com>