mata: audio: Fix speaker phone problems
Speaker phone doesn't work for a very little selection of users,
neither does any video call app. This is because we use voice-speaker-qmic
as the in_snd_device while stock uses voice-speaker-tmic. Fix this so everyone
can enjoy speaker phone.
A big thank you to ThE_MarD for the guidance.

Signed-off-by: celtare21 <celtare21@gmail.com>